refactor(cli): extract sandbox logs helpers
